Create Two way Anova
This code will perform a two-way ANOVA on the data dataset, which contains data on the dependent variable and the independent variables. The assumptions of ANOVA will be checked, and the results of the ANOVA will be interpreted. Post-hoc tests will be performed to compare the means of the groups.